requirements.txt
numpy==2.3.4
openai==2.6.0
openpyxl==3.1.2
pandas==1.5.3
python_dateutil==2.8.2
Requests==2.32.5
sseclient==0.0.27
sseclient_py==1.8.0
streamlit==1.40.1
tqdm==4.66.1
